Extract out materialTable
